# backend/api/Swagger/swagger.json
openapi: 3.0.0
info:
  title: T-Ride
  description: |-
    T-Ride
  version: 0.0.99
...
paths:
...
  /match/driver/invitation:
    post:
      summary: Send Invitation to Passenger
      tags:
        - match
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                driverOrderId:
                  type: integer
                passengerOrderId:
                  type: integer
              required:
                - token
                - driverOrderId
                - passengerOrderId
      responses:
        '200':
          description: Invitation sent successfully
        '401':
          description: Invalid token
        '403':
          description: Not the driver's own order
        '404':
          description: Order not found
        '409':
          description: Invitation already sent, accepted, or order completed
  /match/driver/invitation/{driverOrderId}/{passengerOrderId}:
    delete:
      summary: Cancel Invitation to passengers
      tags:
        - match
      parameters:
        - name: driverOrderId
          in: path
          description: ID of the driver order
          required: true
          schema:
            type: integer
        - name: passengerOrderId
          in: path
          description: ID of the passenger order
          required: true
          schema:
            type: integer
        - name: token
          in: query
          description: Authentication token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invitation canceled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalOrderCount:
                    type: integer
                  abandonCount:
                    type: integer
        '401':
          description: Invalid token
        '403':
          description: Not the driver's own order
        '404':
          description: Order not found
        '409':
          description: Invitation not sent or order completed
  /match/driver/invitation/total/{driverOrderId}:
    get:
      summary: Get Invitations Sent by a Driver Order
      tags:
        - match
      parameters:
        - name: driverOrderId
          in: path
          description: ID of the driver order
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverTotalInvitations'
        '404':
          description: Order not found
  /match/passenger/invitation/total/{passengerOrderId}:
    get:
      summary: Get Driver Orders Inviting a Passenger Order
      tags:
        - match
      parameters:
        - name: passengerOrderId
          in: path
          description: ID of the passenger order
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassengerTotalInvitations'
        '404':
          description: Order not found
  /match/passenger/accepted/{passengerOrderId}:
    get:
      summary: Get the Accepted Driver Order
      tags:
        - match
      parameters:
        - name: passengerOrderId
          in: path
          description: ID of the passenger order
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  driverOrder:
                    $ref: '#/components/schemas/DriverOrderDetails'
        '404':
          description: Order not found
  /match/passenger/invitation/accept:
    post:
      summary: Accept Invitation to a Driver Order
      tags:
        - match
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInvitationRequest'
      responses:
        '200':
          description: Successfully accepted invitation
        '400':
          description: Incorrect parameter format
        '401':
          description: Invalid token
        '403':
          description: Not the passenger's own order
        '404':
          description: Order not found
        '409':
          description: Invitation not received or order already accepted/completed
components:
  schemas:
    DriverOrderDetails:
      type: object
      properties:
        orderId:
          type: integer
        userId:
          type: integer
        startPoint:
          type: object
          properties:
            lng:
              type: number
              example: 120.99777403377531
            lat:
              type: number
              example: 24.787707214648197
        startName:
          type: string
        endPoint:
          type: object
          properties:
            lng:
              type: number
              example: 120.99777403377531
            lat:
              type: number
              example: 24.787707214648197
        endName:
          type: string
        departureTime:
          type: integer
        passengerCount:
          type: integer
    PassengerOrderDetails:
      type: object
      properties:
        orderId:
          type: integer
        userId:
          type: integer
        startPoint:
          type: object
          properties:
            lng:
              type: number
              example: 120.99777403377531
            lat:
              type: number
              example: 24.787707214648197
        startName:
          type: string
        endPoint:
          type: object
          properties:
            lng:
              type: number
              example: 120.99777403377531
            lat:
              type: number
              example: 24.787707214648197
        endName:
          type: string
        pickTime1:
          type: integer
        pickTime2:
          type: integer
        arrivalTime:
          type: integer
        passengerCount:
          type: integer
        fee:
          type: integer
...
    DriverTotalInvitations:
      type: object
      properties:
        invitations:
          type: array
          items:
            type: object
            properties:
              passengerOrder:
                $ref: '#/components/schemas/PassengerOrderDetails'
              accepted:
                type: boolean  
    PassengerTotalInvitations:
      type: object
      properties:
        driverOrders:
          type: array
          items:
            $ref: '#/components/schemas/DriverOrderDetails'
    AcceptInvitationRequest:
      type: object
      properties:
        token:
          type: string
        driverOrderId:
          type: integer
        passengerOrderId:
          type: integer
      required:
        - token
        - driverOrderId
        - passengerOrderId
...